pyproject.toml
[project]
name = "ai-search"
version = "0.1.0"
description = "Candidate Generation & AI Search Pipeline using Azure AI Foundry and Azure AI Search"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "openai>=1.58.0",
    "azure-search-documents>=11.6.0",
    "azure-identity>=1.17.0",
    "azure-ai-inference>=1.0.0b9",
    "aiohttp>=3.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pillow>=10.0",
    "httpx>=0.27",
    "structlog>=24.0",
    "numpy>=1.26",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "types-Pillow>=10.0",
]

[project.scripts]
ai-search-ingest = "ai_search.ingestion.cli:main"
ai-search-index = "ai_search.indexing.cli:main"
ai-search-query = "ai_search.retrieval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_search"]

[tool.uv]
package = true

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow",
]
